1) Introduced a new type of Regard test that uses the JSVGCanvas to    render dynamic content and then grab the screen contents for    comparison. 2) Updated a few script tests to use this new test type. 3) Fixed a minor bug with completely empty SVG Documents. 4) Fixed a minor bug when JSVGCanvas pops down away with the mouse over it.
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/batik/trunk@201196 13f79535-47bb-0310-9956-ffa450edef68

+++ test-sources/org/apache/batik/test/svg/AbstractRenderingAccuracyTest.java
@@ -14,8 +14,9 @@ import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.InputStream;
 import java.io.IOException;
-import java.io.StringWriter;
+import java.io.OutputStream;
 import java.io.PrintWriter;
+import java.io.StringWriter;
 
 import java.net.URL;
 import java.net.MalformedURLException;
@@ -422,7 +423,8 @@ public abstract class AbstractRenderingAccuracyTest extends AbstractTest {
         // encoding failed and we should return that report.
         {
             TestReport encodeTR = encode(svgURL, tmpFileOS);
-            if (encodeTR != null) {
+            if ((encodeTR != null) && 
+                (encodeTR.hasPassed() == false)) {
                 tmpFile.deleteOnExit();
                 return encodeTR;
             }
@@ -625,10 +627,16 @@ public abstract class AbstractRenderingAccuracyTest extends AbstractTest {
         if(!imgFile.exists()){
             imgFile.createNewFile();
         }
+        saveImage(img, new FileOutputStream(imgFile));
+    }
 
+    /**
+     * Saves an image in a given File
+     */
+    protected void saveImage(BufferedImage img, OutputStream os)
+        throws IOException {
         PNGImageEncoder encoder = new PNGImageEncoder
-            (new FileOutputStream(imgFile),
-             PNGEncodeParam.getDefaultEncodeParam(img));
+            (os, PNGEncodeParam.getDefaultEncodeParam(img));
         
         encoder.encode(img);
     }
